+++ src/applications/validate-mhv-account/components/GenericError.jsx
@@ -1,0 +1,27 @@
+import React from 'react';
+import MessageTemplate from './MessageTemplate';
+
+const GenericError = () => {
+  const content = {
+    heading: 'Please try again later',
+    alertContent: (
+      <>
+        <p>
+          We're sorry. Something went wrong on our end, and we couldn't connect
+          you to health tools.
+        </p>
+      </>
+    ),
+    alertStatus: 'error',
+    body: (
+      <>
+        <h3>What you can do:</h3>
+        <p>Please try again later.</p>
+      </>
+    ),
+  };
+
+  return <MessageTemplate content={content} />;
+};
+
+export default GenericError;
