+++ src/applications/validate-mhv-account/validate-mhv-account-entry.jsx
@@ -1,10 +1,11 @@
 import '../../platform/polyfills';
-import React from 'react';
 
 import startApp from '../../platform/startup';
-import ValidateMHVAccount from './ValidateMHVAccount';
+import routes from './routes';
+import manifest from './manifest.json';
 
 startApp({
-  component: <ValidateMHVAccount />,
+  url: manifest.rootUrl,
+  routes,
   entryName: 'my-health-account-validation',
 });
