+++ src/platform/site-wide/mega-menu/mega-menu-link-data-for-authenticated-users.json
@@ -5,6 +5,6 @@
     },
     {
         "title": "My Health",
-        "href": "/my-health/verify/"
+        "href": "/my-health-account-validation/"
     }
 ]
